caching and cli
