+++ environment.yml
@@ -1,8 +1,8 @@
-name: cenv
+name: dropbear
 channels:
   - defaults
 dependencies:
   - pip
   - requests
   - pip:
-    - -r requirements.txt
+      - -r requirements.txt
