--- front/static/css/sortable-theme-bootstrap.css
@@ -1,90 +1,0 @@
-/* line 2, ../sass/_sortable.sass */
-table[data-sortable] {
-  border-collapse: collapse;
-  border-spacing: 0;
-}
-/* line 6, ../sass/_sortable.sass */
-table[data-sortable] th {
-  vertical-align: bottom;
-  font-weight: bold;
-}
-/* line 10, ../sass/_sortable.sass */
-table[data-sortable] th, table[data-sortable] td {
-  text-align: left;
-  padding: 10px;
-}
-/* line 14, ../sass/_sortable.sass */
-table[data-sortable] th:not([data-sortable="false"]) {
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  -o-user-select: none;
-  user-select: none;
-  -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
-  -webkit-touch-callout: none;
-  cursor: pointer;
-}
-/* line 26, ../sass/_sortable.sass */
-table[data-sortable] th:after {
-  content: "";
-  visibility: hidden;
-  display: inline-block;
-  vertical-align: inherit;
-  height: 0;
-  width: 0;
-  border-width: 5px;
-  border-style: solid;
-  border-color: transparent;
-  margin-right: 1px;
-  margin-left: 10px;
-  float: right;
-}
-/* line 40, ../sass/_sortable.sass */
-table[data-sortable] th[data-sorted="true"]:after {
-  visibility: visible;
-}
-/* line 43, ../sass/_sortable.sass */
-table[data-sortable] th[data-sorted-direction="descending"]:after {
-  border-top-color: inherit;
-  margin-top: 8px;
-}
-/* line 47, ../sass/_sortable.sass */
-table[data-sortable] th[data-sorted-direction="ascending"]:after {
-  border-bottom-color: inherit;
-  margin-top: 3px;
-}
-
-/* line 5, ../sass/sortable-theme-bootstrap.sass */
-table[data-sortable].sortable-theme-bootstrap {
-  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-  font-size: 14px;
-  line-height: 20px;
-  color: #333333;
-  background: white;
-}
-/* line 12, ../sass/sortable-theme-bootstrap.sass */
-table[data-sortable].sortable-theme-bootstrap thead th {
-  border-bottom: 2px solid #e0e0e0;
-}
-/* line 15, ../sass/sortable-theme-bootstrap.sass */
-table[data-sortable].sortable-theme-bootstrap tbody td {
-  border-top: 1px solid #e0e0e0;
-}
-/* line 18, ../sass/sortable-theme-bootstrap.sass */
-table[data-sortable].sortable-theme-bootstrap th[data-sorted="true"] {
-  color: #3a87ad;
-  background: #d9edf7;
-  border-bottom-color: #bce8f1;
-}
-/* line 23, ../sass/sortable-theme-bootstrap.sass */
-table[data-sortable].sortable-theme-bootstrap th[data-sorted="true"][data-sorted-direction="descending"]:after {
-  border-top-color: #3a87ad;
-}
-/* line 26, ../sass/sortable-theme-bootstrap.sass */
-table[data-sortable].sortable-theme-bootstrap th[data-sorted="true"][data-sorted-direction="ascending"]:after {
-  border-bottom-color: #3a87ad;
-}
-/* line 31, ../sass/sortable-theme-bootstrap.sass */
-table[data-sortable].sortable-theme-bootstrap.sortable-theme-bootstrap-striped tbody > tr:nth-child(odd) > td {
-  background-color: #f9f9f9;
-}
