--- front/templates/index.html
@@ -1,89 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Asymmetry</title>
-
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"
-    integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
-  <link rel="stylesheet" href="/static/css/sortable-theme-bootstrap.css">
-
-  <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js"
-    integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n"
-    crossorigin="anonymous"></script>
-  <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"
-    integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo"
-    crossorigin="anonymous"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"
-    integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6"
-    crossorigin="anonymous"></script>
-  <script src="/static/js/sortable.min.js"></script>
-  <script src="https://unpkg.com/mustache@latest"></script>
-  <script src="/static/js/asymmetry.js"></script>
-
-</head>
-
-<body>
-
-  <div class="container">
-    <div class="row py-5">
-      <div class="col">
-        <form id="searchForm">
-          <div class="form-row align-items-center">
-            <div class="col">
-              <label class="sr-only" for="authorName">An author's name</label>
-              <input type="text" class="form-control mb-2" id="authorName" placeholder="Enter an author's name">
-            </div>
-            <div class="col-auto">
-              <button type="submit" class="btn btn-primary mb-2">Search</button>
-            </div>
-          </div>
-        </form>
-        <div id="progressText" class="text-secondary font-weight-light"></div>
-      </div>
-    </div>
-
-    <div class="row">
-      <div class="col">
-        <table id="resultsTable" class="table sortable-theme-bootstrap" data-sortable>
-          <thead>
-            <tr>
-              <th>Name</th>
-              <th>Affiliation</th>
-              <th>Number of papers</th>
-              <th>
-                Gender Guess
-                <!-- https://github.com/danklammer/bytesize-icons -->
-                <svg id="i-info" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32" width="12" height="12"
-                  fill="none" stroke="currentcolor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                  data-toggle="popover" data-trigger="hover" data-placement="bottom"
-                  data-content="This column is estimated based on the given name available on ADS.">
-                  <path d=" M16 14 L16 23 M16 8 L16 10" />
-                  <circle cx="16" cy="16" r="14" />
-                </svg>
-              </th>
-            </tr>
-          </thead>
-          <tbody id="tableBody">
-          </tbody>
-          {% raw %}
-          <script id="rowTemplate" type="x-tmpl-mustache">
-            <tr>
-              <td><a href="https://www.google.com/search?query={{ full_name }} astronomy" target="_blank">{{ full_name }}</a></td>
-              <td>{{ most_recent_primary_affiliation }}</td>
-              <td>{{ number_of_articles }}</td>
-              <td>{{ inferred_gender }}</td>
-            </tr>
-          </script>
-          {% endraw %}
-        </table>
-      </div>
-    </div>
-  </div>
-
-</body>
-
-</html>
