+++ src/dropbear/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+__all__ = ["cli"]
+
+from .core import cli
+
+__uri__ = "https://github.com/andycasey/dropbear"
+__author__ = "Andy Casey & Daniel Foreman-Mackey"
+__email__ = "foreman.mackey@gmail.com"
+__license__ = "MIT"
+__description__ = ""
